Refactor: Add multi-database configuration for Flask
- Create migrations
  - Including initial migration and alembic configuration
- Add server/models/base.py file
  - This was made to hold the base class for all models
  it helps with the metadata
- Add server/extensions.py file
  - This is for a cleaner structure of the extensions
  in the flask factory pattern
- Add server/db_initiator.py file
  - This is the renamed db.py file now takes care of the
  databases initialization
- Add server/models/wms_sqlite3_model.py file
  - This was renamed because it clashed with the new models
- Update server/config.py file
  - This was updated to include the new database configurations
- Update server/__init__.py file
  - This was updated to include the new database configurations
- Update server/models/tms_models.py file
  - This is the updated models file to include the new base class
- Update server/models/wms_models.py file
  - This is the updated models file to include the new base class

All in all this was a refactor to include multiple databases in the
application. This was very time consuming :)
