+++ server/__init__.py
@@ -3,7 +3,8 @@ import os
 from flask import Flask
 from flask_cors import CORS
 from .config import Config
-from flask_jwt_extended import JWTManager
+from .extensions import jwt, db, migrate
+
 
 # Create the application instance
 def create_app(config_class=Config):
@@ -17,22 +18,23 @@ def create_app(config_class=Config):
     except OSError:
         pass
 
-    # Logs
-
-
     # Enable CORS
     CORS(app)
 
-    # Initialize JWT
-    jwt = JWTManager()
+    # Initialize the extensions
     jwt.init_app(app)
-
-    # Initialize the databases
-    from . import db
     db.init_app(app)
-    
+    migrate.init_app(app, db)
 
-    # Register blueprints
+    # Register commands
+    from .db_initiator import init_db_command
+    app.cli.add_command(init_db_command)
+
+    with app.app_context():
+        from .models import tms_models
+        from .models import wms_models
+
+    """# Register blueprints
     from .api.errors_handler import errors_blueprint
     from .api.auth import auth_blueprint
     from .api.admin import admin_blueprint
@@ -43,7 +45,7 @@ def create_app(config_class=Config):
     app.register_blueprint(auth_blueprint)
     app.register_blueprint(admin_blueprint)
     app.register_blueprint(onboarding_blueprint)
-    app.register_blueprint(wms_blueprint)
+    app.register_blueprint(wms_blueprint) """
 
     # Return the app
     return app
