+++ server/config.py
@@ -3,6 +3,7 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
+
 class Config:
     """Base configuration"""
     SECRET_KEY = os.getenv('SECRET_KEY')
@@ -10,35 +11,28 @@ class Config:
     JWT_ALGORITHM = os.getenv('JWT_ALGORITHM')
     JWT_ACCESS_TOKEN_EXPIRES = int(os.getenv('JWT_ACCESS_TOKEN_EXPIRES'))
     JWT_REFRESH_TOKEN_EXPIRES = int(os.getenv('JWT_REFRESH_TOKEN_EXPIRES'))
-    TMS_DATABASE_URL = os.getenv('DATABASE_URL')
-    WMS_DATABASE_URL = os.getenv('WMS_DATABASE_URL')
+    SQLALCHEMY_DATABASE_URI = os.getenv('TMS_DATABASE_URL')
+    SQLALCHEMY_BINDS = {
+        'wms': os.getenv('WMS_DATABASE_URL')
+    }
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
 
 
 class Development(Config):
     """Development configuration"""
     DEBUG = True
-    SQLALCHEMY_DATABASE_URI = Config.TMS_DATABASE_URL
-    SQLACLHEMY_BINDS = {
-        "wms": Config.WMS_DATABASE_URL
-    }
 
 
 class Testing(Config):
     """Testing configuration"""
     TESTING = True
     SQLALCHEMY_DATABASE_URI = "sqlite:///:memory:"
-    SQLACLHEMY_BINDS = {
+    SQLALCHEMY_BINDS = {
         "wms": "sqlite:///:memory:"
     }
-    
 
 
 class Production(Config):
     """Production configuration"""
     DEBUG = False
     TESTING = False
-    SQLALCHEMY_DATABASE_URI = Config.TMS_DATABASE_URL
-    SQLACLHEMY_BINDS = {
-        "wms": Config.WMS_DATABASE_URL
-    }
-
