+++ server/models/tms_models.py
@@ -1,8 +1,9 @@
-from ..db import db
-from datetime import datetime
+from ..extensions import db
+from datetime import datetime, timezone
+from .base import Base
 
 
-class Role(db.Model):
+class Role(Base):
     __tablename__ = "roles"
 
     role_id = db.Column(db.Integer, primary_key=True)
@@ -12,7 +13,7 @@ class Role(db.Model):
         return f"Role('{self.role_name}' id: {self.role_id})"
 
 
-class User(db.Model):
+class User(Base):
     __tablename__ = "users"
 
     user_id = db.Column(db.Integer, primary_key=True)
@@ -22,9 +23,9 @@ class User(db.Model):
     role_id = db.Column(db.Integer, db.ForeignKey(
         "roles.role_id"), nullable=False)
     created_at = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     updated_at = db.Column(db.DateTime, default=datetime.now(
-        datetime.timezone.utc), onupdate=datetime.now(datetime.timezone.utc))
+        timezone.utc), onupdate=datetime.now(timezone.utc))
 
     role = db.relationship("Role", backref="users")
 
@@ -32,12 +33,12 @@ class User(db.Model):
         return f"User('{self.username}', '{self.email}')"
 
 
-class UserDetails(db.Model):
+class UserDetails(Base):
     __tablename__ = "user_details"
 
     user_detail_id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey(
-        "users.user_id"), ondelete="CASCADE")
+        "users.user_id", ondelete="CASCADE"))
     first_name = db.Column(db.String(50))
     last_name = db.Column(db.String(50))
     phone_number = db.Column(db.String(15))
@@ -49,15 +50,15 @@ class UserDetails(db.Model):
         return f"UserDetails('{self.first_name}', '{self.last_name}')"
 
 
-class DriverDetails:
+class DriverDetails(Base):
     __tablename__ = "driver_details"
 
     driver_detail_id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey(
-        "users.user_id"), ondelete="CASCADE")
+        "users.user_id", ondelete="CASCADE"))
     license_number = db.Column(db.String(50), unique=True, nullable=False)
     license_expiry = db.Column(db.DateTime, nullable=False)
-    vehicle_id = db.Column(db.Integer, db.ForeginKey(
+    vehicle_id = db.Column(db.Integer, db.ForeignKey(
         "vehicles.vehicle_id"), nullable=False)
     driver_status = db.Column(db.String(50), nullable=False)
 
@@ -68,7 +69,7 @@ class DriverDetails:
         return f"DriverDetails('{self.license_number}', '{self.driver_status}')"
 
 
-class Vehicle(db.Model):
+class Vehicle(Base):
     __tablename__ = "vehicles"
 
     vehicle_id = db.Column(db.Integer, primary_key=True)
@@ -83,7 +84,7 @@ class Vehicle(db.Model):
         return f"Vehicle('{self.vehicle_plate}', '{self.vehicle_type}')"
 
 
-class Warehouse(db.Model):
+class Warehouse(Base):
     __tablename__ = "warehouses"
 
     warehouse_id = db.Column(db.Integer, primary_key=True)
@@ -98,7 +99,7 @@ class Warehouse(db.Model):
         return f"Warehouse('{self.warehouse_name}', '{self.location}')"
 
 
-class Order(db.Model):
+class Order(Base):
     __tablename__ = "orders"
 
     order_id = db.Column(db.Integer, primary_key=True)
@@ -106,9 +107,9 @@ class Order(db.Model):
     customer_id = db.Column(db.Integer, db.ForeignKey("users.user_id"))
     order_status = db.Column(db.String(50), nullable=False)
     created_at = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     updated_at = db.Column(db.DateTime, default=datetime.now(
-        datetime.timezone.utc), onupdate=datetime.now(datetime.timezone.utc))
+        timezone.utc), onupdate=datetime.now(timezone.utc))
 
     customer = db.relationship("User", backref="orders")
 
@@ -116,13 +117,13 @@ class Order(db.Model):
         return f"Order('{self.order_uuid}', '{self.order_status}')"
 
 
-class OrderDetails(db.Model):
+class OrderDetails(Base):
     __tablename__ = "order_details"
 
     order_detail_id = db.Column(db.Integer, primary_key=True)
     order_id = db.Column(db.Integer, db.ForeignKey(
-        "orders.order_id"), ondelete="CASCADE")
-    product_id = db.Column(db.Integer, db.ForeignKey("products.product_id"))
+        "orders.order_id", ondelete="CASCADE"))
+    product_id = db.Column(db.Integer, nullable=False)
     product_name = db.Column(db.String(50), nullable=False)
     priority = db.Column(db.Integer, nullable=False)
     quantity = db.Column(db.Integer, nullable=False)
@@ -135,12 +136,12 @@ class OrderDetails(db.Model):
         return f"OrderDetails('{self.quantity}', '{self.price}')"
 
 
-class Shipment(db.Model):
+class Shipment(Base):
     __tablename__ = "shipments"
 
     shipment_id = db.Column(db.Integer, primary_key=True)
     order_id = db.Column(db.Integer, db.ForeignKey(
-        "orders.order_id"), ondelete="CASCADE")
+        "orders.order_id", ondelete="CASCADE"))
     customer_id = db.Column(db.Integer, db.ForeignKey("users.user_id"))
     driver_id = db.Column(db.Integer, db.ForeignKey("users.user_id"))
     vehicle_id = db.Column(db.Integer, db.ForeignKey("vehicles.vehicle_id"))
@@ -152,9 +153,9 @@ class Shipment(db.Model):
     departure_time = db.Column(db.DateTime)
     arrival_time = db.Column(db.DateTime)
     created_at = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     updated_at = db.Column(db.DateTime, default=datetime.now(
-        datetime.timezone.utc), onupdate=datetime.now(datetime.timezone.utc))
+        timezone.utc), onupdate=datetime.now(timezone.utc))
 
     order = db.relationship("Order", backref="shipments")
     customer = db.relationship("User", backref="shipments")
@@ -166,19 +167,19 @@ class Shipment(db.Model):
         return f"Shipment('{self.status}', '{self.origin}', '{self.destination}')"
 
 
-class Invoice(db.Model):
+class Invoice(Base):
     __tablename__ = "invoices"
 
     invoice_id = db.Column(db.Integer, primary_key=True)
     shipment_id = db.Column(db.Integer, db.ForeignKey(
-        "shipments.shipment_id"), ondelete="CASCADE")
+        "shipments.shipment_id", ondelete="CASCADE"))
     amount = db.Column(db.Float, nullable=False)
     status = db.Column(db.String(50), nullable=False)
     due_date = db.Column(db.DateTime, nullable=False)
     created_at = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     updated_at = db.Column(db.DateTime, default=datetime.now(
-        datetime.timezone.utc), onupdate=datetime.now(datetime.timezone.utc))
+        timezone.utc), onupdate=datetime.now(timezone.utc))
 
     shipment = db.relationship("Shipment", backref="invoices")
 
@@ -186,15 +187,15 @@ class Invoice(db.Model):
         return f"Invoice('{self.amount}', '{self.status}', '{self.due_date}')"
 
 
-class Payment(db.Model):
+class Payment(Base):
     __tablename__ = "payments"
 
     payment_id = db.Column(db.Integer, primary_key=True)
     invoice_id = db.Column(db.Integer, db.ForeignKey(
-        "invoices.invoice_id"), ondelete="CASCADE")
+        "invoices.invoice_id", ondelete="CASCADE"))
     amount = db.Column(db.Float, nullable=False)
     payment_date = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     status = db.Column(db.String(50), nullable=False)
 
     invoice = db.relationship("Invoice", backref="payments")
@@ -203,14 +204,14 @@ class Payment(db.Model):
         return f"Payment('{self.amount}', '{self.status}', '{self.payment_date}')"
 
 
-class AuditLog(db.Model):
+class AuditLog(Base):
     __tablename__ = "audit_logs"
 
     log_id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey("users.user_id"))
     action = db.Column(db.String(50), nullable=False)
     timestamp = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     details = db.Column(db.String(255))
 
     user = db.relationship("User", backref="audit_logs")
@@ -219,14 +220,14 @@ class AuditLog(db.Model):
         return f"AuditLog('{self.action}', '{self.timestamp}')"
 
 
-class Report(db.Model):
+class Report(Base):
     __tablename__ = "reports"
 
     report_id = db.Column(db.Integer, primary_key=True)
     report_type = db.Column(db.String(50), nullable=False)
     generated_by = db.Column(db.Integer, db.ForeignKey("users.user_id"))
     created_at = db.Column(
-        db.DateTime, default=datetime.now(datetime.timezone.utc))
+        db.DateTime, default=datetime.now(timezone.utc))
     data = db.Column(db.String(255))
 
     user = db.relationship("User", backref="reports")
